add cases separately
